EE-16064 Move tests to their correct package
