+++ src/test/java/uk/gov/digital/ho/pttg/api/AuditResourceIntTest.java
@@ -1,4 +1,4 @@
-package uk.gov.digital.ho.pttg;
+package uk.gov.digital.ho.pttg.api;
 
 import org.apache.commons.io.IOUtils;
 import org.json.JSONException;
